+++ packages/apollo-client/src/__tests__/client.ts
@@ -79,7 +79,7 @@ describe('client', () => {
             a
           }
         `,
-      } as any)
+      } as any),
     ).rejects.toThrow(
       'mutation option is required. You must specify your GraphQL document in the mutation option.',
     );
@@ -277,6 +277,78 @@ describe('client', () => {
   });
 
   it('should allow fragments on root query', () => {
+    const query = gql`
+      query {
+        ...QueryFragment
+      }
+
+      fragment QueryFragment on Query {
+        records {
+          id
+          name
+          __typename
+        }
+        __typename
+      }
+    `;
+
+    const data = {
+      records: [
+        { id: 1, name: 'One', __typename: 'Record' },
+        { id: 2, name: 'Two', __typename: 'Record' },
+      ],
+      __typename: 'Query',
+    };
+
+    return clientRoundtrip(query, { data }, null);
+  });
+
+  it('should allow fragments on root query with ifm', () => {
+    const query = gql`
+      query {
+        ...QueryFragment
+      }
+
+      fragment QueryFragment on Query {
+        records {
+          id
+          name
+          __typename
+        }
+        __typename
+      }
+    `;
+
+    const data = {
+      records: [
+        { id: 1, name: 'One', __typename: 'Record' },
+        { id: 2, name: 'Two', __typename: 'Record' },
+      ],
+      __typename: 'Query',
+    };
+
+    const ifm = new IntrospectionFragmentMatcher({
+      introspectionQueryResultData: {
+        __schema: {
+          types: [
+            {
+              kind: 'UNION',
+              name: 'Query',
+              possibleTypes: [
+                {
+                  name: 'Record',
+                },
+              ],
+            },
+          ],
+        },
+      },
+    });
+
+    return clientRoundtrip(query, { data }, null, ifm);
+  });
+
+  it('should merge fragments on root query', () => {
     // The fragment should be used after the selected fields for the query.
     // Otherwise, the results aren't merged.
     // see: https://github.com/apollographql/apollo-client/issues/1479
@@ -1537,15 +1609,19 @@ describe('client', () => {
       cache: new InMemoryCache(),
     });
 
-    const sub1 = client.watchQuery({
-      query: document,
-      variables: variables1,
-    }).subscribe({});
+    const sub1 = client
+      .watchQuery({
+        query: document,
+        variables: variables1,
+      })
+      .subscribe({});
 
-    const sub2 = client.watchQuery({
-      query: document,
-      variables: variables2,
-    }).subscribe({});
+    const sub2 = client
+      .watchQuery({
+        query: document,
+        variables: variables2,
+      })
+      .subscribe({});
 
     sub1.unsubscribe();
     expect(unsubscribed).toBe(false);